Move Team model to a new module
