+++ ansible_base/models/__init__.py
@@ -1,6 +1,7 @@
 from .authenticator import Authenticator
 from .authenticator_map import AuthenticatorMap
-from .organization import AbstractOrganization, AbstractTeam
+from .organization import AbstractOrganization
+from .team import AbstractTeam
 from .social_auth import AuthenticatorUser
 
 __all__ = (
